git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2788 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ cpu/common/textfile.c
@@ -2,7 +2,7 @@
 #include	"strres.h"
 #include	"textfile.h"
 #if defined(SUPPORT_TEXTCNV)
-#include	"textcnv.h"
+#include	"codecnv/textcnv.h"
 #endif
 #if defined(SUPPORT_ARC)
 #include	"arc.h"
@@ -565,7 +565,7 @@ const void		*buf;
 		flushfile(tf);
 		tf->mode = TFMODE_WRITE;
 	}
-	leng = OEMSTRLEN(buffer);
+	leng = (UINT)OEMSTRLEN(buffer);
 #if defined(SUPPORT_TEXTCNV)
 	if (tf->fromoem != NULL) {
 		leng = (tf->fromoem)(tf->cnvbuf, tf->cnvbufsize / tf->width,
